remove internalProcedure
