+++ packages/server/src/core/internals/procedureBuilder.ts
@@ -1,5 +1,7 @@
+import { TRPCError } from '../../error/TRPCError';
+import { getCauseFromUnknown, getErrorFromUnknown } from '../../error/utils';
 import { MaybePromise } from '../../types';
-import { MiddlewareFunction } from '../middleware';
+import { MiddlewareFunction, MiddlewareResult } from '../middleware';
 import { Parser, inferParser } from '../parser';
 import {
   MutationProcedure,
@@ -8,9 +10,11 @@ import {
   QueryProcedure,
   SubscriptionProcedure,
 } from '../procedure';
+import { ProcedureType } from '../types';
 import { RootConfig } from './config';
-import { createInternalBuilder } from './internalProcedure';
-import { ResolveOptions } from './utils';
+import { ParseFn, getParseFn } from './getParseFn';
+import { mergeWithoutOverrides } from './mergeWithoutOverrides';
+import { ResolveOptions, middlewareMarker } from './utils';
 import { DefaultValue as FallbackValue, Overwrite, UnsetMarker } from './utils';
 
 type CreateProcedureReturnInput<
@@ -23,13 +27,8 @@ type CreateProcedureReturnInput<
   _ctx_out: Overwrite<TPrev['_ctx_out'], TNext['_ctx_out']>;
   _input_in: FallbackValue<TNext['_input_in'], TPrev['_input_in']>;
   _input_out: FallbackValue<TNext['_input_out'], TPrev['_input_out']>;
-  _input_parser: FallbackValue<TNext['_input_parser'], TPrev['_input_parser']>;
   _output_in: FallbackValue<TNext['_output_in'], TPrev['_output_in']>;
   _output_out: FallbackValue<TNext['_output_out'], TPrev['_output_out']>;
-  _output_parser: FallbackValue<
-    TNext['_output_parser'],
-    TPrev['_output_parser']
-  >;
 }>;
 
 export interface ProcedureBuilder<TParams extends ProcedureParams> {
@@ -45,10 +44,8 @@ export interface ProcedureBuilder<TParams extends ProcedureParams> {
     _ctx_out: TParams['_ctx_out'];
     _input_in: inferParser<$TParser>['in'];
     _input_out: inferParser<$TParser>['out'];
-    _input_parser: $TParser;
     _output_in: TParams['_output_in'];
     _output_out: TParams['_output_out'];
-    _output_parser: TParams['_output_parser'];
   }>;
   /**
    * Add an output parser to the procedure.
@@ -62,10 +59,8 @@ export interface ProcedureBuilder<TParams extends ProcedureParams> {
     _ctx_out: TParams['_ctx_out'];
     _input_in: TParams['_input_in'];
     _input_out: TParams['_input_out'];
-    _input_parser: TParams['_input_parser'];
     _output_in: inferParser<$TParser>['in'];
     _output_out: inferParser<$TParser>['out'];
-    _output_parser: $TParser;
   }>;
   /**
    * Add a meta data to the procedure.
@@ -79,10 +74,8 @@ export interface ProcedureBuilder<TParams extends ProcedureParams> {
     _ctx_out: TParams['_ctx_out'];
     _input_in: TParams['_input_in'];
     _input_out: TParams['_input_out'];
-    _input_parser: TParams['_input_parser'];
     _output_in: TParams['_output_in'];
     _output_out: TParams['_output_out'];
-    _output_parser: TParams['_output_parser'];
   }>;
   /**
    * Add a middleware to the procedure.
@@ -94,7 +87,7 @@ export interface ProcedureBuilder<TParams extends ProcedureParams> {
    * Extend the procedure with another procedure.
    * @warning The TypeScript inference fails when chaining concatenated procedures.
    */
-  unstable_concat<$ProcedureReturnInput extends ProcedureBuilder<any>>(
+  unstable_concat<$ProcedureReturnInput extends AnyProcedureBuilder>(
     proc: $ProcedureReturnInput,
   ): $ProcedureReturnInput extends ProcedureBuilder<infer $TParams>
     ? CreateProcedureReturnInput<TParams, $TParams>
@@ -116,24 +109,7 @@ export interface ProcedureBuilder<TParams extends ProcedureParams> {
         >
       >
     : Procedure<TParams>;
-  /**
-   * Query procedure
-   */
-  query<$TOutput>(
-    resolver: (
-      opts: ResolveOptions<TParams>,
-    ) => MaybePromise<FallbackValue<TParams['_output_in'], $TOutput>>,
-  ): UnsetMarker extends TParams['_output_out']
-    ? QueryProcedure<
-        Overwrite<
-          TParams,
-          {
-            _output_in: $TOutput;
-            _output_out: $TOutput;
-          }
-        >
-      >
-    : QueryProcedure<TParams>;
+
   /**
    * Query procedure
    */
@@ -190,19 +166,264 @@ export interface ProcedureBuilder<TParams extends ProcedureParams> {
         >
       >
     : SubscriptionProcedure<TParams>;
+  /**
+   * @internal
+   */
+  _def: {
+    input?: Parser;
+    output?: Parser;
+    meta?: Record<string, unknown>;
+    resolver?: ProcedureBuilderResolver;
+    middlewares: ProcedureBuilderMiddleware[];
+    mutation?: boolean;
+    query?: boolean;
+    subscription?: boolean;
+  };
 }
 
-export function createBuilder<TConfig extends RootConfig>(): ProcedureBuilder<{
+type AnyProcedureBuilder = ProcedureBuilder<any>;
+export type ProcedureBuilderDef = AnyProcedureBuilder['_def'];
+
+export type ProcedureBuilderMiddleware = MiddlewareFunction<any, any>;
+
+export type ProcedureBuilderResolver = (
+  opts: ResolveOptions<any>,
+) => Promise<unknown>;
+
+function createNewBuilder(
+  def1: ProcedureBuilderDef,
+  def2: Partial<ProcedureBuilderDef>,
+) {
+  const { middlewares = [], ...rest } = def2;
+
+  // TODO: maybe have a fn here to warn about calls
+  return createBuilder({
+    ...mergeWithoutOverrides(def1, rest),
+    middlewares: [...def1.middlewares, ...middlewares],
+  } as any);
+}
+
+export function createBuilder<TConfig extends RootConfig>(
+  initDef?: ProcedureBuilderDef,
+): ProcedureBuilder<{
   _config: TConfig;
   _ctx_in: TConfig['ctx'];
   _ctx_out: TConfig['ctx'];
   _input_in: UnsetMarker;
   _input_out: UnsetMarker;
-  _input_parser: UnsetMarker;
   _output_in: UnsetMarker;
   _output_out: UnsetMarker;
-  _output_parser: UnsetMarker;
   _meta: TConfig['meta'];
 }> {
-  return createInternalBuilder() as any;
+  const _def: ProcedureBuilderDef = initDef || {
+    middlewares: [],
+  };
+
+  return {
+    _def,
+    input(input) {
+      const parser = getParseFn(input);
+      return createNewBuilder(_def, {
+        input,
+        middlewares: [createInputMiddleware(parser)],
+      }) as AnyProcedureBuilder;
+    },
+    output(output: Parser) {
+      const parseOutput = getParseFn(output);
+      return createNewBuilder(_def, {
+        output,
+        middlewares: [createOutputMiddleware(parseOutput)],
+      }) as AnyProcedureBuilder;
+    },
+    meta(meta) {
+      return createNewBuilder(_def, {
+        meta: meta as Record<string, unknown>,
+      }) as AnyProcedureBuilder;
+    },
+    /** @deprecated **/
+    resolve(resolver) {
+      return createResolver(_def, resolver);
+    },
+    unstable_concat(builder) {
+      return createNewBuilder(_def, builder._def) as any;
+    },
+    use(middleware) {
+      return createNewBuilder(_def, {
+        middlewares: [middleware],
+      }) as AnyProcedureBuilder;
+    },
+    query(resolver) {
+      return createResolver(
+        { ..._def, query: true },
+        resolver,
+      ) as QueryProcedure<any>;
+    },
+    mutation(resolver) {
+      return createResolver(
+        { ..._def, mutation: true },
+        resolver,
+      ) as MutationProcedure<any>;
+    },
+    subscription(resolver) {
+      return createResolver(
+        { ..._def, subscription: true },
+        resolver,
+      ) as SubscriptionProcedure<any>;
+    },
+  };
+}
+
+function createInputMiddleware<T>(
+  parse: ParseFn<T>,
+): ProcedureBuilderMiddleware {
+  return async function inputMiddleware({ next, rawInput }) {
+    let input: ReturnType<typeof parse>;
+    try {
+      input = await parse(rawInput);
+    } catch (cause) {
+      throw new TRPCError({
+        code: 'BAD_REQUEST',
+        cause: getCauseFromUnknown(cause),
+      });
+    }
+    // TODO fix this typing?
+    return next({ input } as any);
+  };
+}
+
+function createOutputMiddleware<T>(
+  parse: ParseFn<T>,
+): ProcedureBuilderMiddleware {
+  return async function outputMiddleware({ next }) {
+    const result = await next();
+    if (!result.ok) {
+      // pass through failures without validating
+      return result;
+    }
+    try {
+      const data = await parse(result.data);
+      return {
+        ...result,
+        data,
+      };
+    } catch (cause) {
+      throw new TRPCError({
+        message: 'Output validation failed',
+        code: 'INTERNAL_SERVER_ERROR',
+        cause: getCauseFromUnknown(cause),
+      });
+    }
+  };
+}
+
+function createResolver(
+  _def: ProcedureBuilderDef,
+  resolver: (opts: ResolveOptions<any>) => MaybePromise<any>,
+) {
+  const finalBuilder = createNewBuilder(_def, {
+    resolver,
+    middlewares: [
+      async function resolveMiddleware(opts) {
+        const data = await resolver(opts);
+        return {
+          marker: middlewareMarker,
+          ok: true,
+          data,
+          ctx: opts.ctx,
+        } as const;
+      },
+    ],
+  });
+
+  return createProcedureCaller(finalBuilder._def);
+}
+
+/**
+ * @internal
+ */
+export interface ProcedureCallOptions {
+  ctx: unknown;
+  rawInput: unknown;
+  input?: unknown;
+  path: string;
+  type: ProcedureType;
+}
+
+const codeblock = `
+If you want to call this function on the server, you do the following:
+This is a client-only function.
+
+const caller = appRouter.createCaller({
+  /* ... your context */
+});
+
+const result = await caller.call('myProcedure', input);
+`.trim();
+
+function createProcedureCaller(_def: ProcedureBuilderDef): Procedure<any> {
+  const procedure = async function resolve(opts: ProcedureCallOptions) {
+    if (!opts || !('rawInput' in opts)) {
+      throw new Error(codeblock);
+    }
+
+    // run the middlewares recursively with the resolver as the last one
+    const callRecursive = async (
+      callOpts: { ctx: any; index: number; input?: unknown } = {
+        index: 0,
+        ctx: opts.ctx,
+      },
+    ): Promise<MiddlewareResult<any>> => {
+      try {
+        const middleware = _def.middlewares[callOpts.index];
+        const result = await middleware({
+          ctx: callOpts.ctx,
+          type: opts.type,
+          path: opts.path,
+          rawInput: opts.rawInput,
+          meta: _def.meta,
+          input: callOpts.input,
+          next: async (nextOpts?: { ctx: any; input?: any }) => {
+            return await callRecursive({
+              index: callOpts.index + 1,
+              ctx:
+                nextOpts && 'ctx' in nextOpts
+                  ? { ...callOpts.ctx, ...nextOpts.ctx }
+                  : callOpts.ctx,
+              input:
+                nextOpts && 'input' in nextOpts
+                  ? nextOpts.input
+                  : callOpts.input,
+            });
+          },
+        });
+        return result;
+      } catch (cause) {
+        return {
+          ok: false,
+          error: getErrorFromUnknown(cause),
+          marker: middlewareMarker,
+        };
+      }
+    };
+
+    // there's always at least one "next" since we wrap this.resolver in a middleware
+    const result = await callRecursive();
+
+    if (!result) {
+      throw new TRPCError({
+        code: 'INTERNAL_SERVER_ERROR',
+        message:
+          'No result from middlewares - did you forget to `return next()`?',
+      });
+    }
+    if (!result.ok) {
+      // re-throw original error
+      throw result.error;
+    }
+    return result.data;
+  };
+  procedure._def = _def;
+  procedure.meta = _def.meta;
+
+  return procedure as Procedure<any>;
 }
